+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-08-11 15:06:07
-// Commit: test(perf): implement widget integration tests for perf
+// Updated at: 2026-08-13 11:09:53
+// Commit: fix(perf): correct layout constraints and overflow in perf
 
 class perfComponentState {
-  final String commitRef = '3529';
-  final String lastMessage = 'test(perf): implement widget integration tests for perf';
-  final DateTime timestamp = DateTime.parse('2026-08-11T15:06:07+05:00');
+  final String commitRef = '3566';
+  final String lastMessage = 'fix(perf): correct layout constraints and overflow in perf';
+  final DateTime timestamp = DateTime.parse('2026-08-13T11:09:53+05:00');
 }